water sensor

+++ water_sensor/water_sensor_readout_V0.py
@@ -42,7 +42,7 @@ while True:
 
 		# print('Raw reading from Arduino :' + str(valueRead)) # Read the newest output from the Arduino
 		y_vals = valueRead.decode().strip().split(',')  # s1, s2, s3
-		y_vals = [str(int(y)) for y in y_vals]
+		y_vals = [str(float(y)) for y in y_vals]
 		print(y_vals)
